requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlmodel>=0.0.8
psycopg2-binary>=2.9.6
passlib>=1.7.4
python-jose>=3.3.0
bcrypt>=4.0.1
python-multipart>=0.0.6
pydantic>=1.10.7